+++ frontend/src/components/Dashboard.js
@@ -7,7 +7,7 @@ import UserContext from "../context/UserContext";
 import MyGames from "./dashboardComponents/MyGames";
 import GameShop from "./dashboardComponents/GameShop";
 
-export default function Dashboard() {
+export default function Dashboard({userGames}) {
   //state for antall spill i MyGames
   const [numGames, setNumGames] = useState(0);
 
@@ -61,7 +61,7 @@ export default function Dashboard() {
                 My Games
               </Link>
             </section>
-            <MyGames />
+            <MyGames userGames={userGames}/>
           </section>
           <section className="gs">
             <section className="gs-txt-box">
