+++ frontend/src/components/dashboardComponents/MyGames.js
@@ -1,44 +1,15 @@
 import GameCard from "../GameCard"
-import UserContext from '../../context/UserContext';
-import { useState, useEffect, useContext } from "react";
-import { fetchGamesByUsername} from "../../utils/sanity/userServices";
 
 //Komponent som henter 
-export default function MyGames(){
-
-  const [userGames, setUserGames] = useState([]);
-  const { user } = useContext(UserContext);
-
-
-   async function getUserGames(username) {
-    const data = await fetchGamesByUsername(username);
-    const userGames = data.games
-    if(userGames.length > 4){
-      return userGames.slice(0,4)
-    }
-    return userGames
-  } 
-
-   useEffect(() => {
-    if (user) {
-      getUserGames(user.username)
-        .then((userGames) => {
-         
-          setUserGames(userGames);
-        })
-        .catch((error) => {
-          console.error(error);
-          window.location.href = "/login";
-        })
-    }
-  }, [user]);
+export default function MyGames({userGames}){
+  const games = userGames.slice(0,4)
 
     return (
       <>
-      {userGames !== undefined ? ( 
+      {games !== undefined ? ( 
         <article className="gs-box">
       
-          {userGames?.map((item, index) => (
+          {games?.map((item, index) => (
             <GameCard
               key={index}
               title={item.game.title}
